New StateSystem
State System based on the concept of Flags. Only a start. It should be possible for the entities to be in a few states at once. Instead of checking if the entity is in a particular state, the game should check if the given 'substate' is part of the state. ^= used to change binary data and toggle 'substates'.  GivingBirth, Eating, Drinking and Mating shouldn't go along with other 'substates' and be toggled on and off, but be set as the main state, as being in them prevents the entity from doing anything else (except for dying or fleeing). Work in progress so old code is commented out.