Bool states in State Data
Added bool variables to the State Data that can be used to easily tell what 'substates' are currently toggled on/off. Their presence added a useful panel to the inspector as well. Fleeing is followed by the pregnancy check to add an option for the entity to start giving birth while being chased - bad luck if that happens I guess.